fix: (will revert, only for future reference) Fix existing bug that has existed for years: Always use up to date methods for searching with string lookup
But, this fix now makes the string cache lookup slower than not using the cache (up to 50% slower patching time). But, the string lookup seems no longer needed since strings literals are now instruction filters and thus there is no second iterative search of the method to match strings. So the performance gain for doing string caching (even with the old broken logic) is only around 3% performance improvement. So, the cache is no longer needed.